requirements.txt
pandas==1.5.3
numpy==1.22.4
PyYAML==6.0
scikit-learn==1.2.2
seaborn==0.12.2
scipy==1.10.1
scikit-plot==0.3.7
matplotlib==3.7.1
plotly==5.13.1
lightgbm==3.3.5
xgboost==1.7.6
shap==0.42.0
requests==2.31.0                  